projects/006-firestore-bqbackfill/requirements.txt
firebase-admin>=6.5.0
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.26.0
pandas>=2.2.0
pyarrow>=20.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0